# Optimize performance of keys definition