--- a/lib/definition/version.rb
+++ b/lib/definition/version.rb
@@ -1,5 +1,5 @@
 # frozen_string_literal: true
 
 module Definition
-  VERSION = "1.0.0"
+  VERSION = "1.1.0"
 end